% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{build_visnetwork}
\alias{build_visnetwork}
\title{Build a visNetwork widget from a StateGraph}
\usage{
build_visnetwork(graph)
}
\arguments{
\item{graph}{A \link{StateGraph} object.}
}
\value{
A \code{visNetwork} htmlwidget.
}
\description{
Build a visNetwork widget from a StateGraph
}
\keyword{internal}
