% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpointer.R
\name{Checkpointer}
\alias{Checkpointer}
\title{Checkpointer base class}
\value{
\verb{list(step = <int>, state = <list>)} or \code{NULL}.

State snapshot list or \code{NULL}.

Character vector.
}
\description{
Abstract interface for persisting workflow state. Concrete subclasses:
\link{MemoryCheckpointer}, \link{RDSCheckpointer}, \link{SQLiteCheckpointer}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Checkpointer-save}{\code{Checkpointer$save()}}
\item \href{#method-Checkpointer-load_latest}{\code{Checkpointer$load_latest()}}
\item \href{#method-Checkpointer-load_step}{\code{Checkpointer$load_step()}}
\item \href{#method-Checkpointer-list_threads}{\code{Checkpointer$list_threads()}}
\item \href{#method-Checkpointer-clone}{\code{Checkpointer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Checkpointer-save"></a>}}
\if{latex}{\out{\hypertarget{method-Checkpointer-save}{}}}
\subsection{Method \code{save()}}{
Save a state snapshot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Checkpointer$save(thread_id, step, state_snapshot)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{Character. Identifies the workflow run.}

\item{\code{step}}{Integer. Current step number.}

\item{\code{state_snapshot}}{Named list from \code{WorkflowState$snapshot()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Checkpointer-load_latest"></a>}}
\if{latex}{\out{\hypertarget{method-Checkpointer-load_latest}{}}}
\subsection{Method \code{load_latest()}}{
Load the most recent snapshot for a thread.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Checkpointer$load_latest(thread_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{Character.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Checkpointer-load_step"></a>}}
\if{latex}{\out{\hypertarget{method-Checkpointer-load_step}{}}}
\subsection{Method \code{load_step()}}{
Load a snapshot at an exact step.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Checkpointer$load_step(thread_id, step)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{Character.}

\item{\code{step}}{Integer.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Checkpointer-list_threads"></a>}}
\if{latex}{\out{\hypertarget{method-Checkpointer-list_threads}{}}}
\subsection{Method \code{list_threads()}}{
List all thread IDs that have checkpoints.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Checkpointer$list_threads()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Checkpointer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Checkpointer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Checkpointer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
