% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termination.R
\name{compose_termination}
\alias{compose_termination}
\alias{|.termination_condition}
\alias{&.termination_condition}
\title{Compose termination conditions}
\usage{
\method{|}{termination_condition}(a, b)

\method{&}{termination_condition}(a, b)
}
\arguments{
\item{a, b}{\code{termination_condition} objects.}
}
\value{
A composite \code{termination_condition} object.
}
\description{
Combine two termination conditions with \code{|} (OR) or \code{&} (AND).
}
\examples{
cond <- max_turns(10L) | cost_limit(1.0)
cond2 <- max_turns(10L) & text_match("done", channel = "status")
}
