% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\docType{data}
\name{END}
\alias{END}
\title{Sentinel: end of graph}
\format{
An object of class \code{puppeteer_sentinel} of length 1.
}
\usage{
END
}
\description{
Use \code{END} as the \code{to} argument of \code{\link[=add_edge]{add_edge()}} or in a \code{route_map} to
indicate that execution should stop.
}
\keyword{datasets}
