% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termination.R
\name{max_turns}
\alias{max_turns}
\title{Termination condition: maximum iterations}
\usage{
max_turns(n)
}
\arguments{
\item{n}{Integer. Maximum number of node executions.}
}
\value{
A \code{termination_condition} S3 object.
}
\description{
Stops execution when the iteration counter reaches \code{n}.
}
\examples{
cond <- max_turns(10L)
check_termination(cond, NULL, 10L, 0)
}
