% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpointer.R
\name{memory_checkpointer}
\alias{memory_checkpointer}
\title{Create an in-memory checkpointer}
\usage{
memory_checkpointer()
}
\value{
A \link{MemoryCheckpointer} object.
}
\description{
Create an in-memory checkpointer
}
\examples{
cp <- memory_checkpointer()
}
