% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{reducer_append}
\alias{reducer_append}
\title{Reducer: append new value to a list}
\usage{
reducer_append()
}
\value{
A two-argument function \verb{function(old, new)}.
}
\description{
Wraps \code{new} in a \code{list()} and concatenates it to \code{old}. Useful for
accumulating messages.
}
\examples{
r <- reducer_append()
r(list("a"), "b")
}
