% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{reducer_merge}
\alias{reducer_merge}
\title{Reducer: merge lists with \code{modifyList}}
\usage{
reducer_merge()
}
\value{
A two-argument function \verb{function(old, new)}.
}
\description{
Performs a shallow merge of \code{new} into \code{old} using \code{\link[=modifyList]{modifyList()}}.
Useful for nested configuration state.
}
\examples{
r <- reducer_merge()
r(list(a = 1, b = 2), list(b = 99, c = 3))
}
