% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpointer.R
\name{sqlite_checkpointer}
\alias{sqlite_checkpointer}
\title{Create a SQLite checkpointer}
\usage{
sqlite_checkpointer(path)
}
\arguments{
\item{path}{Character. Path to the SQLite database file.}
}
\value{
A \link{SQLiteCheckpointer} object.
}
\description{
Create a SQLite checkpointer
}
\examples{
\dontrun{
cp <- sqlite_checkpointer(tempfile(fileext = ".sqlite"))
}
}
