% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\docType{data}
\name{START}
\alias{START}
\title{Sentinel: start of graph}
\format{
An object of class \code{puppeteer_sentinel} of length 1.
}
\usage{
START
}
\description{
Use \code{START} as the \code{from} argument of \code{\link[=add_edge]{add_edge()}} to define the graph
entry point.
}
\keyword{datasets}
