% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{state_graph}
\alias{state_graph}
\title{Create a StateGraph}
\usage{
state_graph(state_schema)
}
\arguments{
\item{state_schema}{A \link{WorkflowState} or a named list of channel specs.}
}
\value{
A \link{StateGraph} object.
}
\description{
Create a StateGraph
}
\examples{
schema <- workflow_state(result = list(default = NULL))
g <- state_graph(schema)
}
