% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{visualize_graph}
\alias{visualize_graph}
\title{Render a StateGraph visualization}
\usage{
visualize_graph(graph, engine)
}
\arguments{
\item{graph}{A \link{StateGraph} object.}

\item{engine}{Character. One of \code{"dot"}, \code{"visnetwork"}, \code{"mermaid"}.}
}
\value{
A widget or prints to console.
}
\description{
Render a StateGraph visualization
}
\keyword{internal}
