% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpointer.R
\name{rds_checkpointer}
\alias{rds_checkpointer}
\title{Create an RDS file checkpointer}
\usage{
rds_checkpointer(path)
}
\arguments{
\item{path}{Character. Directory to store \code{.rds} checkpoint files.}
}
\value{
An \link{RDSCheckpointer} object.
}
\description{
Create an RDS file checkpointer
}
\examples{
\dontrun{
cp <- rds_checkpointer(tempdir())
}
}
